webserver/requirements.txt
redis
python-multipart
aiofiles
jinja2
numpy
Pillow